Create Car Vet - 2018 ECNA Regional Contest
Problem A of the 2018 ECNA Regional Contest, that was part of the ICPC (International Collegiate Programming Contest) North America Regional Contests, specifically the East Central North America (ECNA) region. This contest is one of the regional competitions held to determine which teams qualify for the ICPC World Finals.

The ECNA Regional Contest includes universities from several U.S. states, primarily in the Midwest, such as Ohio, Indiana, Pennsylvania, Michigan, and Kentucky, among others. In the contest, teams of three students are given a set of algorithmic problems to solve in a limited time, usually five hours. They use languages like C++, Java, and Python to solve as many problems as they can, with the goal of achieving the highest number of correct solutions in the shortest amount of time.